# Conflicts:
#	ruoyi-admin/src/main/resources/application-druid.yml
#	ruoyi-admin/src/main/resources/application.yml
#	ruoyi-admin/src/main/resources/i18n/messages.properties
#	ruoyi-admin/src/main/resources/mybatis/mybatis-config.xml

+++ ruoyi-system/src/main/java/com/ruoyi/system/mapper/SysRoleMapper.java
@@ -39,7 +39,7 @@ public interface SysRoleMapper
      * @param userId 用户ID
      * @return 选中角色ID列表
      */
-    public List<Integer> selectRoleListByUserId(Long userId);
+    public List<Long> selectRoleListByUserId(Long userId);
 
     /**
      * 通过角色ID查询角色
